+++ MOLA/GUIs/TreeLab/__init__.py
@@ -785,9 +785,9 @@ class MainWindow(QMainWindow):
 
             elif NewValue[0].isdigit():
                 if '.' in NewValue:
-                    NewValue = np.array([NewValue],dtype=np.float,order='F')
+                    NewValue = np.array([NewValue],dtype=np.float64,order='F')
                 else:
-                    NewValue = np.array([NewValue],dtype=np.int,order='F')
+                    NewValue = np.array([NewValue],dtype=np.int32,order='F')
 
             
             indices = self.tree.selectionModel().selectedIndexes()
@@ -1362,9 +1362,9 @@ class MainWindow(QMainWindow):
                     node_cgns.setValue(None)
                 elif new_value[0].isdigit():
                     if '.' in new_value:
-                        new_value = np.array([new_value],dtype=np.float,order='F')
+                        new_value = np.array([new_value],dtype=np.float64,order='F')
                     else:
-                        new_value = np.array([new_value],dtype=np.int,order='F')
+                        new_value = np.array([new_value],dtype=np.int32,order='F')
                 node_cgns.setValue(new_value)
 
             elif isinstance(value,list):
